backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pymysql
pandas
python-dotenv
joblib
scikit-learn
google-generativeai
apscheduler
httpx